Move sources over to new tck-inspection
