+++ tools/tck-inspection/src/test/scala/org/opencypher/tools/tck/inspection/CountScenariosTest.scala
@@ -25,7 +25,7 @@
  * described as "implementation extensions to Cypher" or as "proposed changes to
  * Cypher that are not yet approved by the openCypher community".
  */
-package org.opencypher.tools.tck.reporting
+package org.opencypher.tools.tck.inspection
 
 import java.net.URI
 import java.util
